feat : 날짜 체크 기능 모듈로 분리
schedule.js 를 간결하게 만들기 위해 checkDate.js를 모듈로 분리
